Trait arcu::Rcu

pub trait Rcu {
    type Item;
    type Pool: EpochCounterPool;

    // Required methods
    fn new(
        initial: impl Into<Arc<Self::Item>>,
        epoch_counter_pool: Self::Pool,
    ) -> Self;
    fn replace(&self, new_value: impl Into<Arc<Self::Item>>) -> Arc<Self::Item>;
    unsafe fn raw_read(&self, epoch_counter: &EpochCounter) -> Arc<Self::Item>;
    unsafe fn raw_try_update(
        &self,
        update: impl FnMut(&Self::Item) -> Option<Arc<Self::Item>>,
        epoch_counter: &EpochCounter,
    ) -> Option<Arc<Self::Item>>;

    // Provided methods
    fn read(&self) -> RcuRef<Self::Item, Self::Item>
       where Self: Rcu<Pool = GlobalEpochCounterPool> { ... }
    fn try_update<F, R>(&self, update: F) -> Option<Arc<Self::Item>>
       where Self: Rcu<Pool = GlobalEpochCounterPool>,
             F: FnMut(&Self::Item) -> Option<R>,
             R: Into<Arc<Self::Item>> { ... }
}

Required Associated Types

type Item

type Pool: EpochCounterPool

Required Methods

fn new( initial: impl Into<Arc<Self::Item>>, epoch_counter_pool: Self::Pool, ) -> Self

fn replace(&self, new_value: impl Into<Arc<Self::Item>>) -> Arc<Self::Item>

Replace the Rcu’s content with a new value

This does not synchronize writes and the last to update the active_value pointer wins.

all writes that do not win will be lost, though not leaked. This will block until the old value can be reclaimed, i.e. all threads witnessed to be in the read critical sections have been witnessed to have left the critical section at least once

unsafe fn raw_read(&self, epoch_counter: &EpochCounter) -> Arc<Self::Item>

Safety

The epoch counter must not be used concurrently

The epoch counter must belong to the EpochCounterPool of this Rcu

unsafe fn raw_try_update( &self, update: impl FnMut(&Self::Item) -> Option<Arc<Self::Item>>, epoch_counter: &EpochCounter, ) -> Option<Arc<Self::Item>>

Update the Rcu using the provided update function Retries when the Rcu has been updated/replaced between reading the old value and writing the new value Aborts when the update function returns None

Safety

The epoch counter must not be used concurrently

The epoch counter must belong to the EpochCounterPool of this Rcu

Provided Methods

fn read(&self) -> RcuRef<Self::Item, Self::Item>
where Self: Rcu<Pool = GlobalEpochCounterPool>,

Read the value of the Rcu for the current epoch

Blocking

The initial read on each thread may block while registering the epoch counter. Further read on the same thread won’t block even for different Rcu.

Procedure

Register the Epoch Counter (only done once per thread, may block)

atomically increment the epoch counter (by one from even to odd)

atomically load the arc pointer

atomically increment the arc strong count

atomically increment the epoch counter (by one from odd back to even)

fn try_update<F, R>(&self, update: F) -> Option<Arc<Self::Item>>
where Self: Rcu<Pool = GlobalEpochCounterPool>, F: FnMut(&Self::Item) -> Option<R>, R: Into<Arc<Self::Item>>,

Update the Rcu using the provided update function Retries when the Rcu has been updated/replaced between reading the old value and writing the new value Aborts when the update function returns None

Object Safety

This trait is not object safe.